api-server/requirements.txt
requests==2.24.0
starlette==0.13.6
pydantic==1.6.1
fastapi==0.60.1
uvicorn==0.11.6
SQLAlchemy==1.3.18
psycopg2-binary==2.8.5
alembic==1.4.2
bcrypt==3.1.7
PyJWT==1.7.1
